+++ tests/test_load_template.py
@@ -105,6 +105,46 @@ class FileSystemLoaderTestCase(unittest.TestCase):
 
             self.assertEqual(template.tree, updated_template.tree)
 
+    def test_template_cache_size(self):
+        """Test that we can control the template cache size."""
+        with tempfile.TemporaryDirectory() as tmpdir:
+            template_path = Path(tmpdir) / "somefile.txt"
+            another_template_path = Path(tmpdir) / "otherfile.txt"
+
+            # Initial template content
+            with template_path.open("w") as fd:
+                fd.write("hello there\n")
+
+            with another_template_path.open("w") as fd:
+                fd.write("goodbye there\n")
+
+            # Cache size of zero sets auto_reload to False
+            env = Environment(
+                loader=FileSystemLoader(search_path=tmpdir),
+                cache_size=0,
+            )
+            self.assertFalse(env.auto_reload)
+
+            # Very small cache size.
+            env = Environment(
+                loader=FileSystemLoader(search_path=tmpdir),
+                cache_size=1,
+            )
+
+            template = env.get_template(name=str(template_path))
+            self.assertTrue(template.is_up_to_date)
+
+            same_template = env.get_template(name=str(template_path))
+            self.assertTrue(template.is_up_to_date)
+
+            # Cached tempalate is the same object
+            self.assertEqual(template.tree, same_template.tree)
+
+            # Push the first template out of the cache
+            another_template = env.get_template(name=str(another_template_path))
+            self.assertTrue(another_template.is_up_to_date)
+            self.assertEqual(len(env.cache), 1)
+
     def test_template_not_found(self):
         """Test that we get an error if the template does not exist."""
         env = Environment(loader=FileSystemLoader(search_path="tests/fixtures/"))